requirements.txt
absl-py==2.2.2
asgiref==3.9.1
astunparse==1.6.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
fonttools==4.57.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
keras==3.3.3
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
namex==0.0.8
numpy==2.1.3
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pillow==11.1.0
protobuf==5.29.4
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
requests==2.32.3
rich==14.0.0
scikit-learn==1.6.1
scipy==1.15.2
setuptools==78.1.0
six==1.16.0
sqlparse==0.5.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==3.0.1
threadpoolctl==3.6.0
typing_extensions==4.13.1
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.46.1
wrapt==1.17.2